fix: resolve all 70 TODO issues across 5 modules
Session 2 fixes (2026-06-21):
- HIGH: ChatMessage playerId int→Long + V5 BIGINT migration (H-NEW-7);
  LuaJ sandbox hardened with instruction limits (H-NEW-16)
- MEDIUM Common: StatsRegistry resettable (M-NEW-1); GameAPI event
  dispatcher wiring (M-NEW-13); chat_messages BIGINT migration (M-NEW-18);
  production queue right-click cancel (M-NEW-23); key binding modifier
  capture (M-NEW-24)
- MEDIUM Core: Deterministic collections for lockstep (M-NEW-8,9,10);
  AttackMove autoEngage flag (M-NEW-11); Patrol return origin (M-NEW-12)
- MEDIUM Server: RateLimitFilter XFF trust (M-NEW-16); WS size limits
  (M-NEW-17); pagination (M-NEW-19); GlobalExceptionHandler (M-NEW-20);
  lobby cleanup (M-NEW-21); SessionService overwrite warning (M-NEW-22)
- MEDIUM Web: API health endpoint (M-NEW-25); demo data banners (M-NEW-26);
  matchmaking server-first (M-NEW-27); Prisma FK relations (M-NEW-28)
- LOW: 18 polish issues fixed (L-NEW-1 through L-NEW-22)
- Updated TODO.md: 70/79 fixed, 3 false positives, 5 deferred, 0 open

5 items deferred (alpha scope): H-NEW-9, H-NEW-11, H-NEW-12,
H-NEW-13, H-NEW-14 (build protocol, map loading, campaign, audio,
GameSession persistence)

+++ aow2-server/src/main/java/com/aow2/server/model/ChatMessage.java
@@ -26,7 +26,7 @@ public class ChatMessage {
     private String matchId;
 
     @Column(name = "player_id", nullable = false)
-    private int playerId;
+    private Long playerId;
 
     @Column(name = "message", nullable = false, length = 500)
     private String message;
@@ -48,7 +48,7 @@ public class ChatMessage {
      * @param playerId the player ID who sent the message
      * @param message  the chat message content
      */
-    public ChatMessage(String matchId, int playerId, String message) {
+    public ChatMessage(String matchId, Long playerId, String message) {
         this.matchId = matchId;
         this.playerId = playerId;
         this.message = message;
@@ -61,8 +61,8 @@ public class ChatMessage {
     public String getMatchId() { return matchId; }
     public void setMatchId(String matchId) { this.matchId = matchId; }
 
-    public int getPlayerId() { return playerId; }
-    public void setPlayerId(int playerId) { this.playerId = playerId; }
+    public Long getPlayerId() { return playerId; }
+    public void setPlayerId(Long playerId) { this.playerId = playerId; }
 
     public String getMessage() { return message; }
     public void setMessage(String message) { this.message = message; }
